YD-398 Split the crypto package in two * seckey with secret key encryption * pubkey with public key encryption
