+++ appservice/src/main/java/nu/yona/server/analysis/rest/BuddyActivityController.java
@@ -36,7 +36,7 @@ import nu.yona.server.analysis.service.DayActivityOverviewDto;
 import nu.yona.server.analysis.service.PostPutActivityCommentMessageDto;
 import nu.yona.server.analysis.service.WeekActivityDto;
 import nu.yona.server.analysis.service.WeekActivityOverviewDto;
-import nu.yona.server.crypto.CryptoSession;
+import nu.yona.server.crypto.seckey.CryptoSession;
 import nu.yona.server.messaging.rest.MessageController;
 import nu.yona.server.messaging.service.MessageDto;
 import nu.yona.server.subscriptions.rest.BuddyController;
