+++ appservice/src/main/java/nu/yona/server/analysis/rest/UserActivityController.java
@@ -38,9 +38,9 @@ import nu.yona.server.analysis.service.DayActivityDto;
 import nu.yona.server.analysis.service.DayActivityOverviewDto;
 import nu.yona.server.analysis.service.DayActivityWithBuddiesDto;
 import nu.yona.server.analysis.service.DayActivityWithBuddiesDto.ActivityForOneUser;
+import nu.yona.server.crypto.seckey.CryptoSession;
 import nu.yona.server.analysis.service.WeekActivityDto;
 import nu.yona.server.analysis.service.WeekActivityOverviewDto;
-import nu.yona.server.crypto.CryptoSession;
 import nu.yona.server.goals.rest.ActivityCategoryController;
 import nu.yona.server.goals.rest.GoalController;
 import nu.yona.server.messaging.service.MessageDto;
