+++ core/src/main/java/nu/yona/server/crypto/pubkey/PublicKeyDecryptor.java
@@ -2,12 +2,16 @@
  * Copyright (c) 2015, 2016 Stichting Yona Foundation This Source Code Form is subject to the terms of the Mozilla Public License,
  * v. 2.0. If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
  *******************************************************************************/
-package nu.yona.server.crypto;
+package nu.yona.server.crypto.pubkey;
 
 import java.security.GeneralSecurityException;
 import java.security.PrivateKey;
 
 import javax.crypto.Cipher;
+
+import nu.yona.server.crypto.CryptoException;
+import nu.yona.server.crypto.seckey.CryptoSession;
+import nu.yona.server.crypto.seckey.SecretKeyUtil;
 
 public class PublicKeyDecryptor implements Decryptor
 {
@@ -39,7 +43,7 @@ public class PublicKeyDecryptor implements Decryptor
 			Cipher decryptCipher = Cipher.getInstance(PublicKeyUtil.CIPHER_TYPE);
 			decryptCipher.init(Cipher.DECRYPT_MODE, privateKey);
 
-			return CryptoUtil.decrypt(PublicKeyUtil.CURRENT_SMALL_PLAINTEXT_CRYPTO_VARIANT_NUMBER, decryptCipher, ciphertext);
+			return SecretKeyUtil.decrypt(PublicKeyUtil.CURRENT_SMALL_PLAINTEXT_CRYPTO_VARIANT_NUMBER, decryptCipher, ciphertext);
 		}
 		catch (GeneralSecurityException e)
 		{
