+++ core/src/main/java/nu/yona/server/crypto/seckey/ByteFieldEncrypter.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2015, 2016 Stichting Yona Foundation This Source Code Form is subject to the terms of the Mozilla Public License,
  * v. 2.0. If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
  *******************************************************************************/
-package nu.yona.server.crypto;
+package nu.yona.server.crypto.seckey;
 
 import java.util.Base64;
 
@@ -10,16 +10,21 @@ import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
 @Converter
-public class StringFieldEncrypter implements AttributeConverter<String, String>
+public class ByteFieldEncrypter implements AttributeConverter<byte[], String>
 {
 	@Override
-	public String convertToDatabaseColumn(String attribute)
+	public String convertToDatabaseColumn(byte[] plaintext)
 	{
-		return (attribute == null) ? null : Base64.getEncoder().encodeToString(CryptoUtil.encryptString(attribute));
+		if (plaintext == null)
+		{
+			return null;
+		}
+
+		return Base64.getEncoder().encodeToString(SecretKeyUtil.encryptBytes(plaintext));
 	}
 
 	@Override
-	public String convertToEntityAttribute(String dbData)
+	public byte[] convertToEntityAttribute(String dbData)
 	{
 		try
 		{
@@ -28,7 +33,7 @@ public class StringFieldEncrypter implements AttributeConverter<String, String>
 				return null;
 			}
 
-			return CryptoUtil.decryptString(Base64.getDecoder().decode(dbData));
+			return SecretKeyUtil.decryptBytes(Base64.getDecoder().decode(dbData));
 		}
 		catch (RuntimeException ex)
 		{
