+++ core/src/main/java/nu/yona/server/crypto/seckey/StringFieldEncrypter.java
@@ -2,25 +2,24 @@
  * Copyright (c) 2015, 2016 Stichting Yona Foundation This Source Code Form is subject to the terms of the Mozilla Public License,
  * v. 2.0. If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
  *******************************************************************************/
-package nu.yona.server.crypto;
+package nu.yona.server.crypto.seckey;
 
 import java.util.Base64;
-import java.util.UUID;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
 @Converter
-public class UUIDFieldEncrypter implements AttributeConverter<UUID, String>
+public class StringFieldEncrypter implements AttributeConverter<String, String>
 {
 	@Override
-	public String convertToDatabaseColumn(UUID attribute)
+	public String convertToDatabaseColumn(String attribute)
 	{
-		return (attribute == null) ? null : Base64.getEncoder().encodeToString(CryptoUtil.encryptUuid(attribute));
+		return (attribute == null) ? null : Base64.getEncoder().encodeToString(SecretKeyUtil.encryptString(attribute));
 	}
 
 	@Override
-	public UUID convertToEntityAttribute(String dbData)
+	public String convertToEntityAttribute(String dbData)
 	{
 		try
 		{
@@ -29,7 +28,7 @@ public class UUIDFieldEncrypter implements AttributeConverter<UUID, String>
 				return null;
 			}
 
-			return CryptoUtil.decryptUuid(Base64.getDecoder().decode(dbData));
+			return SecretKeyUtil.decryptString(Base64.getDecoder().decode(dbData));
 		}
 		catch (RuntimeException ex)
 		{
