+++ core/src/main/java/nu/yona/server/crypto/seckey/UUIDFieldEncrypter.java
@@ -2,24 +2,25 @@
  * Copyright (c) 2015, 2016 Stichting Yona Foundation This Source Code Form is subject to the terms of the Mozilla Public License,
  * v. 2.0. If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
  *******************************************************************************/
-package nu.yona.server.crypto;
+package nu.yona.server.crypto.seckey;
 
 import java.util.Base64;
+import java.util.UUID;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
 @Converter
-public class LongFieldEncrypter implements AttributeConverter<Long, String>
+public class UUIDFieldEncrypter implements AttributeConverter<UUID, String>
 {
 	@Override
-	public String convertToDatabaseColumn(Long attribute)
+	public String convertToDatabaseColumn(UUID attribute)
 	{
-		return (attribute == null) ? null : Base64.getEncoder().encodeToString(CryptoUtil.encryptLong(attribute));
+		return (attribute == null) ? null : Base64.getEncoder().encodeToString(SecretKeyUtil.encryptUuid(attribute));
 	}
 
 	@Override
-	public Long convertToEntityAttribute(String dbData)
+	public UUID convertToEntityAttribute(String dbData)
 	{
 		try
 		{
@@ -28,7 +29,7 @@ public class LongFieldEncrypter implements AttributeConverter<Long, String>
 				return null;
 			}
 
-			return CryptoUtil.decryptLong(Base64.getDecoder().decode(dbData));
+			return SecretKeyUtil.decryptUuid(Base64.getDecoder().decode(dbData));
 		}
 		catch (RuntimeException ex)
 		{
